Rename "TwoX 128 hash" to  "TwoX 64 Concat"
I spent a long time trying to figure out what was I doing wrong, because my hashes didn't match with the hashes of the docs.

It turns out that when the docs were written [the XXH128 hash](https://github.com/Cyan4973/xxHash) didn't exist yet, and that's probably why back then we used to refer to the ["TwoX 64 Concat" hash](https://docs.substrate.io/v3/runtime/storage/#common-substrate-hashers) as "TwoX 128" hash. However, now that the 128 version of the TwoX hash exists, we should make sure that we always use the term "TwoX 64 Concat", otherwise it can be confusing.