Update sp-io cargo.toml dependency branch version
With the latest node template, if you paste the sp-io dependecy into pallet/template/cargo.toml you will get "failed to load manifest for dependency" error. Incrementing the branch version from .17 to .18 fixed the issue for me. I have documented the fix here: https://stackoverflow.com/questions/70403605/failed-to-load-manifest-for-dependency/71786336#71786336

This fix will improve the tutorial experience, although it was a good learning experience for me to fix this issue :)