auto-select local node in portal
The Polkadot/Substrate Portal defaults to Polkadot the first time the portal is used. An instruction step is missing in the `Authorize access for the third node` section to switch to the local node as per the screenshot included in the documentation beneath it. This proposed change would auto select the local node, but it may be better to add a step on how to manually switch.