Update broken pallet-nicks dependency (#1726)
* Update broken pallet-nicks dependency

The current way the dependency is written does not work with the current substrate-node-template.

In the auto update to polkadot-v0.9.34, speech marks were added around the branch part of this command for the other pallets, but was not updated in this tutorial. We get this error without the speech marks included.

error: failed to select a version for `signature`.

With them included it works fine though. 

One thing I would add is that I get a lot of warnings with the speech marks added around branch on every pallet, if I remove all the speech marks in every Cargo.toml file, I get no warnings. It works either way if all the versions of branch are in sync. I'm not a Substrate expert so I'm not sure the reasoning behind this, but my update fixes the dependency.

* Update content/md/en/docs/tutorials/work-with-pallets/add-a-pallet.md

Co-authored-by: lisa-parity <92225469+lisa-parity@users.noreply.github.com>